LASB-2105 Rename CAA schema sub folder to common
