+++ crime-applications-adaptor/src/main/java/uk/gov/justice/laa/crime/applications/adaptor/mapper/crimeapply/CrimeApplyMapper.java
@@ -4,9 +4,9 @@ import org.apache.commons.lang3.EnumUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Component;
 import uk.gov.justice.laa.crime.applications.adaptor.model.crimeapplicationsadaptor.CrimeApplication;
-import uk.gov.justice.laa.crime.applications.adaptor.model.crimeapplicationsadaptor.InterestOfJustice;
-import uk.gov.justice.laa.crime.applications.adaptor.model.crimeapplicationsadaptor.MagistrateCourt;
-import uk.gov.justice.laa.crime.applications.adaptor.model.crimeapplicationsadaptor.Supplier;
+import uk.gov.justice.laa.crime.applications.adaptor.model.crimeapplicationsadaptor.common.InterestOfJustice;
+import uk.gov.justice.laa.crime.applications.adaptor.model.crimeapplicationsadaptor.common.MagistrateCourt;
+import uk.gov.justice.laa.crime.applications.adaptor.model.crimeapplicationsadaptor.common.Supplier;
 import uk.gov.justice.laa.crime.applications.adaptor.model.criminalapplicationsdatastore.MaatApplication;
 import uk.gov.justice.laa.crime.applications.adaptor.model.criminalapplicationsdatastore.general.Ioj;
 import uk.gov.justice.laa.crime.applications.adaptor.model.criminalapplicationsdatastore.general.Provider;
