+++ crime-applications-adaptor/src/main/resources/schemas/crimeapplicationsadaptor/common/applicant.json
