+++ crime-applications-adaptor/src/main/resources/schemas/crimeapplicationsadaptor/common/caseDetails.json
